CoachAI/requirements.txt
streamlit==1.31.1
pydantic==2.6.1
python-dotenv==1.0.1 